+++ app/src/main/java/com/example/brebner/swarmthing/ChallengeProtectedChecker.java
@@ -23,7 +23,7 @@ public class ChallengeProtectedChecker extends ChallengeChecker {
 
     public String validate(Recorder recorder) {
         SharedPreferences sharedPreferences = PreferenceManager.getDefaultSharedPreferences(recorder.getContext());
-        if (! timeMatches(sharedPreferences.getInt(recorder.getContext().getString(R.string.other_maxtime_key), 0))) {
+        if (! timeMatches(sharedPreferences.getLong(recorder.getContext().getString(R.string.other_maxtime_key), 0))) {
             return "time does not match";
         }
         deaths = 0;
